fix wrong datatype
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@62901 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/linux/universal/linux-6.12/fs/ntfsplus/ntfs_iomap.c
@@ -679,7 +679,7 @@ const struct iomap_ops ntfs_dio_iomap_ops = {
 	.iomap_end		= ntfs_write_iomap_end,
 };
 
-static ssize_t ntfs_map_blocks(struct iomap_writepage_ctx *wpc,
+static int ntfs_map_blocks(struct iomap_writepage_ctx *wpc,
 		struct inode *inode, loff_t offset, unsigned int len)
 {
 	if (offset < wpc->iomap.offset ||
